backend/requirements.txt
anyio==4.8.0
asgiref==3.8.1
blinker==1.9.0
click==8.1.8
diff-match-patch==20241021
dj-database-url==2.3.0
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==24.3
django-import-export==4.3.7
django-phonenumber-field==8.0.0
django-unfold==0.50.0
djangorestframework==3.15.2
djangorestframework-gis==1.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.3.1
et_xmlfile==2.0.0
Flask==3.1.0
GDAL==3.10.2
geos==0.2.3
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
import_export==0.3.1
Jinja2==3.1.6
lxml==5.3.1
MarkupSafe==3.0.2
openpyxl==3.1.5
packaging==24.2
phonenumbers==9.0.0
pillow==11.1.0
psycopg2-binary==2.9.10
pyjwt==2.10.1
sniffio==1.3.1
sqlparse==0.5.3
supabase==2.14.0
tablib==3.8.0
typing_extensions==4.12.2
tzdata==2025.1
watchfiles==1.0.4
Werkzeug==3.1.3
whitenoise==6.9.0